refactor(xtrellis): extract client|server subcommands to files
