+++ go/0kn/cmd/xtrellis/server.go
@@ -1,0 +1,38 @@
+package main
+
+import (
+	"log"
+
+	"github.com/31333337/bmrng/go/trellis/config"
+	"github.com/31333337/bmrng/go/trellis/errors"
+	"github.com/31333337/bmrng/go/trellis/network"
+	"github.com/31333337/bmrng/go/trellis/server"
+)
+
+// from trellis/cmd/server/
+func LaunchServer(args ArgsServer) {
+	serversFile := args.ServerFile
+	groupsFile := args.GroupFile
+	addr := args.Addr
+	errors.Addr = addr
+
+	log.Printf("Launching server with address %s", addr)
+
+	servers, err := config.UnmarshalServersFromFile(serversFile)
+	if err != nil {
+		log.Fatalf("Could not read servers file %s", serversFile)
+	}
+
+	groups, err := config.UnmarshalGroupsFromFile(groupsFile)
+	if err != nil {
+		log.Fatalf("Could not read group file %s", groupsFile)
+	}
+
+	// will start in blocked state
+	h := server.NewHandler()
+	server := server.NewServer(&config.Servers{Servers: servers}, &config.Groups{Groups: groups}, h, addr)
+
+	server.TcpConnections.LaunchAccepts()
+	network.RunServer(h, server, servers, addr)
+	config.Flush()
+}
